web/backend/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
